Move String and other sizes to v_int64

+++ src/oatpp/network/Connection.cpp
@@ -55,7 +55,7 @@ data::v_io_size Connection::write(const void *buff, data::v_io_size count){
 
 #if defined(WIN32) || defined(_WIN32)
 
-  auto result = ::send(m_handle, (const char*) buff, (size_t)count, 0);
+  auto result = ::send(m_handle, (const char*) buff, (int)count, 0);
 
   if(result == SOCKET_ERROR) {
 
@@ -106,7 +106,7 @@ data::v_io_size Connection::read(void *buff, data::v_io_size count){
 
 #if defined(WIN32) || defined(_WIN32)
 
-  auto result = ::recv(m_handle, (char*)buff, (size_t)count, 0);
+  auto result = ::recv(m_handle, (char*)buff, (int)count, 0);
 
   if(result == SOCKET_ERROR) {
